Selecciona y suma
Al tocar una carta el GameManager se queda con el resultado de la suma.
