+++ Assets/Script/Card.cs
@@ -6,12 +6,17 @@ using UnityEngine.EventSystems;
 
 public class Card : MonoBehaviour
 {
-    public int id; //lugar que ocupara la carta y forma para identificarla
+    public int id; //lugar que ocupara la carta y forma para identificarla  
+
+    public int result;
+
     public int value; //Sera el valor de la Carta   
     private Transform target; // Posicion que buscara segun su id
     public float speed = 0.5f; //velocidad para llegar a su posicion
     public bool esta = false; // si la carta esta o no en la mesa
     public Text valueText; // canvas del valor de la carta
+
+    
 
     private void Start()
     {
@@ -31,5 +36,12 @@ public class Card : MonoBehaviour
     {
         id = Id;
         value = Value;
+    }
+
+    private void OnMouseDown()
+    {
+        GameManager.gameManager.result = value + GameManager.gameManager.cardsOnTable[6].value; ;
+        Destroy(GameManager.gameManager.cardsOnTable[6].gameObject);
+        Destroy(gameObject);
     }
 }
